Start over in my HSoM Journey
Did Chapter 1 again. I decided to spice things up by learning
optparse-applicative in the process, so there's now a relatively
full-featured commandline app

Chapter 1 used to be Chapter 0 - I was working out of the old
pdf from Yale, but now I bought the book.

So I moved the old `ch0` code to
`ch1`, and the old `ch1` and `ch2` code are now in `old` as an
archive; I'll delete them as I go through the corresponding
chapters of the book.