Add a preliminary EuterpeaExports module
The idea is to re-export a decent-size subset of the
Euterpea namespace, to the point where it would be reasonable
to explicitly import or use an import qualified for most other
functions.

I think this is necessary because Euterpea exports an enourmous
number of names, including a through g, which causes excessive
name clashes. Even more so since many of the names (note, chord,
etc) overlap with names you'd naturally want to give locals when
doing musical composition.